requirements.txt
alembic==1.7.6
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
coverage==6.3.2
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.7
pytest==7.0.1
python-decouple==3.6
six==1.16.0
SQLAlchemy==1.4.31
tomli==2.0.1
Werkzeug==2.0.2
WTForms==3.0.1
